Assistant checkpoint: Update back button navigation in LearnDetailView
Assistant generated file changes:
- client/src/pages/LearnDetailView.tsx: Fix back button navigation

---

User prompt:

The back-button on the @client/src/pages/LearnDetailView.tsx still isn't working on mobile and desktop. There is no way for me to go back.

Can you look in the dependency what type of 'videoplayer' we're using so to see if there is a function that we  can use that allows us to go 'back'